+++ CHANGELOG.md
@@ -14,8 +14,11 @@ Changes are grouped as follows
 
 
 ## [0.28.0] - 2020-10-27
-### Added
+### Changed
 - Renamed match_from and match_to in entity matcher, as well as timestamp->time in contextualization jobs.
+- Entity matching matches renamed to true_matches
+### Added
+- Entity matching pipelines rejected_matches and score_threshold options
 
 ## [0.27.0] - 2020-10-20
 ### Added